+++ lib/pure/math.nim
@@ -68,8 +68,14 @@ when defined(c) or defined(cpp):
   proc c_signbit(x: SomeFloat): cint {.importc: "signbit", header: "<math.h>".}
 
   func c_frexp*(x: cfloat, exponent: var cint): cfloat {.
-      importc: "frexpf", header: "<math.h>".}
+      importc: "frexpf", header: "<math.h>", deprecated: "Use `frexp` instead".}
   func c_frexp*(x: cdouble, exponent: var cint): cdouble {.
+      importc: "frexp", header: "<math.h>", deprecated: "Use `frexp` instead".}
+
+  # don't export `c_frexp` in the future and remove `c_frexp2`.
+  func c_frexp2(x: cfloat, exponent: var cint): cfloat {.
+      importc: "frexpf", header: "<math.h>".}
+  func c_frexp2(x: cdouble, exponent: var cint): cdouble {.
       importc: "frexp", header: "<math.h>".}
 
 func binom*(n, k: int): int =
@@ -1000,7 +1006,7 @@ func frexp*[T: float32|float64](x: T, exponent: var int): T =
 
   when not defined(js):
     var exp: cint
-    result = c_frexp(x, exp)
+    result = c_frexp2(x, exp)
     exponent = exp
   else:
     if x == 0.0: